+++ asf_search/__init__.py
@@ -1,5 +1,4 @@
-# backport of importlib.metadata for python < 3.8
-from importlib_metadata import PackageNotFoundError, version
+from importlib.metadata import PackageNotFoundError, version
 
 ## Setup logging now, so it's available if __version__ fails:
 import logging